+++ user_frontend/src/main.tsx
@@ -2,11 +2,20 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from  './App'
 import './index.css'
+import { ClerkProvider } from '@clerk/clerk-react'
 
 
+
+const PUBLISHABLE_KEY = import.meta.env.VITE_CLERK_PUBLISHABLE_KEY
+
+if (!PUBLISHABLE_KEY) {
+  throw new Error('Add your Clerk publishable key to the .env file')
+}
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
+    <ClerkProvider publishableKey={PUBLISHABLE_KEY} afterSignOutUrl="/">
       <App />
-  </StrictMode>
+    </ClerkProvider>
+  </StrictMode>,
 )
